run on real device

+++ app/src/main/java/com/example/catchmeifyoucan/ui/runs/RunsModel.kt
@@ -8,5 +8,5 @@ data class RunsModel(
     @field:Json(name = "start_lng") var start_lng: Double = 0.0,
     @field:Json(name = "end_lat") var end_lat: Double = 0.0,
     @field:Json(name = "end_lng") var end_lng: Double = 0.0,
-    @field:Json(name ="time") val time: Long
+    @field:Json(name ="time") val time: Double
 )
